app/requirements.txt
Flask==3.1.2
ipython>=8.29.0
pymysql==1.1.2
mysql-connector-python==9.4.0
pytest==8.3.4